requirements.txt
requests
beautifulsoup4
gitpython